# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface ConversationRenameInput {
  title: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  language: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
  language: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  language: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export interface CortexConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface CortexMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CortexConversationInput {
  title: string;
}

export interface CortexMessageInput {
  content: string;
}

export interface CortexConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: CortexMessage[];
}